feat: ignore output_path option
When LMPROF_ENABLE_IO is disabled the library still expects an output_path
argument (often nil) for API consistency. Having the output_path as the final
argument leads to ambiguities due to the profile mode strings being variadic.

This commit introduces a preprocessor flag to disable/skip the handling of any
output_path argument. The flag only receives light mention in the README and the
onus will be on any derivative works to update the documentation to reflect
their compilation configuration.